Made DB and vault related changes

+++ Models/DTOs/cartItemInfo.cs
@@ -3,6 +3,6 @@
     public class cartItemInfo
     {
         public product? product{ get; set; }
-        public float quantity { get; set; }
+        public double quantity { get; set; }
     }
 }
