+++ Models/DTOs/newCartItem.cs
@@ -3,6 +3,6 @@
     public class newCartItem
     {
         public int product_id { get; set; }
-        public float quantity { get; set; }
+        public double quantity { get; set; }
     }
 }
